# Conflicts:
#	command/genesis/params.go
#	e2e-polybft/framework/test-cluster.go
#	helper/common/common.go
#	state/executor.go
#	types/rlp_encoding_test.go
#	types/rlp_unmarshal.go

+++ network/server_identity.go
@@ -112,7 +112,7 @@ func (s *Server) setupIdentity() error {
 	identityService := identity.NewIdentityService(
 		s,
 		s.logger,
-		int64(s.config.Chain.Params.ChainID),
+		s.config.Chain.Params.ChainID,
 		s.host.ID(),
 	)
